dump/requirements.txt
pillow
requests
pandas
exifread
PyYAML
python-ulid
tqdm

# Note: exiftool must be installed separately
# brew install exiftool